+++ src/Dracodeum/Kit/Traits/Properties/Arrayable.php
@@ -7,8 +7,6 @@
 
 namespace Dracodeum\Kit\Traits\Properties;
 
-use Dracodeum\Kit\Utilities\Data as UData;
-
 /**
  * This trait implements the <code>Dracodeum\Kit\Interfaces\Arrayable</code> interface 
  * when the properties trait is used.
@@ -19,17 +17,8 @@ trait Arrayable
 {
 	//Implemented final public methods (Dracodeum\Kit\Interfaces\Arrayable)
 	/** {@inheritdoc} */
-	final public function toArray(bool $recursive = false): array
+	final public function toArray(): array
 	{
-		$array = $this->getAll();
-		if ($recursive) {
-			foreach ($array as &$value) {
-				if (is_object($value)) {
-					UData::evaluate($value, null, false, false, true);
-				}
-			}
-			unset($value);
-		}
-		return $array;
+		return $this->getAll();
 	}
 }