+++ changes.md
@@ -5,3 +5,5 @@
   * `Vendor::isLibrary()` &#8594; `System::isFramework()`
 * Renamed `Stringifiable` interface and trait to `Stringable`
   * `non_stringifiable` &#8594; `non_stringable` (property)
+* Removed `$recursive` parameter from most methods
+  * `Arrayable::toArray(bool $recursive = false)` &#8594; `Arrayable::toArray()`